[Messages] mwValidNameMessageBody missing parameters print in message
This is a bit kludgy and can probably be broken out into a helper method
of some sort that is more elegant. If we do this more than once we
should probably consider doing that.

Now if parameters are missing, it returns the messages:
Missing required information - name
Missing required information - message
Missing required information - name + message

respectively